Refactored code
- Create new files to handle conf files.
