+++ src/queue_handler.c
@@ -9,7 +9,7 @@
 #include "slog.h"
 #include "http_handler.h"
 #include "utils.h"
-#include "config.h"
+#include "conf_handler.h"
 
 #include "queue_handler.h"
 #include "resource_handler.h"
